Refactor SettingView into SettingsScreen
